refactor, start module base just a bit
