+++ src/main/java/me/ryguy/ctfbot/modules/Module.java
@@ -1,0 +1,4 @@
+package me.ryguy.ctfbot.modules;
+
+public abstract class Module {
+}
