+++ src/main/java/me/ryguy/ctfbot/modules/ModuleCommand.java
@@ -1,0 +1,11 @@
+package me.ryguy.ctfbot.modules;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface ModuleCommand {
+}
