+++ src/main/java/me/ryguy/ctfbot/modules/Modules.java
@@ -1,0 +1,4 @@
+package me.ryguy.ctfbot.modules;
+
+public enum Modules {
+}
